requirements.txt
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.2.0
beautifulsoup4==4.9.3
certifi==2021.5.30
charset-normalizer==2.0.4
colorama==0.4.4
coverage==5.5
Django==3.2.6
django-database-view==0.2.1
django-endless-pagination==2.0
django-extensions==3.1.3
django-filter==21.1
djangorestframework==3.12.4
freezegun==1.1.0
furl==2.1.2
gunicorn==20.1.0
idna==3.2
iniconfig==1.1.1
jsonschema==3.2.0
numpy==1.21.2
orderedmultidict==1.0.1
packaging==21.0
pandas==1.3.2
pluggy==1.0.0
pook==1.0.2
psycopg2-binary==2.9.1
py==1.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
pytest==6.2.5
pytest-cov==2.12.1
pytest-django==4.4.0
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
rollbar==0.16.2
selenium==3.141.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.6
xmltodict==0.12.0
